+++ lib/analyseData.js
@@ -14,22 +14,19 @@ const getWordsWithCountPerTweet = (tweetText, wordsWithCounters) => {
   return clonedWords;
 };
 
-/* We need all words in caps lock and without any non alphabetic chars. */
-const cleanUpTextArray = (textArray) => {
-  const removedSigns = textArray.map(tweet => tweet.replace(/[^\w\s\'']/g, ' '));
-  const upperCased = removedSigns.map(tweet => tweet.toUpperCase());
-  const removedNonWordLetters = upperCased.map((tweet) => {
-    const splited = tweet.split(' ');
-    const removedNonWords = splited.filter((str) => {
-      if (str.length > 1) return str;
-      return (str === 'A' || str === 'I');
-    });
-
-    return ` ${removedNonWords.join(' ')}`; // We need whitespace at the beginning
+const cleanUpTextTweet = (tweetText) => {
+  const removedSigns = tweetText.replace(/[^\w\s\'']/g, ' ');
+  const upperCased = removedSigns.toUpperCase();
+  const splited = upperCased.split(' ');
+  const removedNonWords = splited.filter((str) => {
+    if (str.length > 1) return str;
+    return (str === 'A' || str === 'I');
   });
-
-  return removedNonWordLetters;
+  return ` ${removedNonWords.join(' ')}`;
 };
+
+/* We need all words in caps lock and without any non alphabetic chars. */
+const cleanUpTextArray = textArray => textArray.map(tweetText => cleanUpTextTweet(tweetText));
 
 const analyseText = (tweetsTextArray, words) => {
   let wordsWithCounters = words.map(word => ({ word, count: 0 }));
@@ -45,9 +42,24 @@ const analyseText = (tweetsTextArray, words) => {
   return sortDesc(wordsWithCounters);
 };
 
+const getTweetsContainingTheWord = (words, tweetsObjectsArray) => {
+  const temp = tweetsObjectsArray.map((tweet) => {
+    const tweetClone = Object.assign(tweet, { words: {} });
+    const cleanedTweetText = cleanUpTextTweet(tweet.full_text);
+
+    for (let i = 0; i < words.length; i += 1) {
+      if (cleanedTweetText.includes(` ${words[i]} `)) tweetClone.words.word = words[i];
+    }
+
+    return tweetClone;
+  });
+
+  const filtered = temp.filter(tweet => Object.prototype.hasOwnProperty.call(tweet.words, 'word'));
+  return filtered;
+};
+
 module.exports = {
   analyseText,
   extractTextProperties,
-  /* temp */
-  cleanUpTextArray,
+  getTweetsContainingTheWord,
 };
