+++ sass/graph.scss
@@ -3,7 +3,7 @@
 
 @import '../node_modules/bulma/sass/elements/form';
 
-@import 'test';
+@import 'correlationGraph';
 
 @import 'modal';
