--- sass/test.scss
@@ -1,11 +1,0 @@
-.xAxis, .yAxis {
-  text {
-    fill: white
-  };
-
-  path {
-    fill: none;
-    stroke: white;
-    stroke-width: 1px;
-  }
-}
