--- views/modal.pug
@@ -1,16 +1,0 @@
-div.modal#edit-words-list
-        div.modal-background
-        div.modal-card
-            header.modal-card-head
-                div.field.has-addons
-                    div.control
-                        input.input#input-word(type='text' placeholder='Enter new word')
-                    div.control
-                        button.button.is-info#add-word Add
-            section.modal-card-body
-                div.background-overlay
-                div.content
-
-            footer.modal-card-foot
-                a.button.is-primary#submit(disabled) Save new list
-                a.button#cancel I changed my mind
