--- views/test.pug
@@ -1,12 +1,0 @@
-extends layout
-
-block content
-
-    script(src='https://d3js.org/d3.v4.js')
-    script.
-        var tweets = !{tweets}
-        var words = !{words}
-
-    div.main-graph-container
-
-    script(src='/test.js')
